requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
gunicorn==21.2.0
idna==3.4
packaging==23.1
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.5
whitenoise==6.5.0
